requirements.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
cachetools==4.2.1
camelot-py==0.8.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.13
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.19.0
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.7.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.0
oauth2==1.9.0.post1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.5.1.48
openpyxl==3.0.6
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pdfminer.six==20201018
pickleshare==0.7.5
Pillow==8.1.0
pipreq==0.4
pipreqs==0.4.10
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.6.0
Pygments==2.7.4
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.2
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.5
sortedcontainers==2.3.0
streamlit==0.76.0
terminado==0.9.2
testpath==0.4.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
validators==0.18.2
virtualenv==20.4.2
watchdog==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
